app/requirements.txt
flask
flask-swagger-ui
Flask-PyMongo
flask-injector
flask_cors
python-dotenv
